Cs {StreamMetabolism}	R Documentation

Saturation Concentration at temp

Description

Calculates the concentration(mg/L) @ 100

Usage

Cs(x)

Arguments

x
Temperature in Degrees Celcius

Details

enter one temperature or a zoo time series of temperature

Value

single value or time series of mg/L@saturation for that temperature

Author(s)

Stephen A Sefick Jr.

References

APHA (2005). Standard Methods for the Examination of Water and Wastewater. 21st Edition. Eds: Eaton, A.D., L.S. Clesceri, E.W. Rice and A. E. Greenberg. Washington D.C., published jointly by the American Public Health Association, the American Water Works Association and the Water Environment Federation.

Examples

#single temperature
temp <- sample(20:30, 1)
Cs(temp)

#USGS Data (DOTemp)
library(chron)
library(zoo)
data(DOTemp)
Cs(DOTemp[,1])

[Package StreamMetabolism version 1.1.2 Index]