DOTemp {StreamMetabolism}	R Documentation

Rhode River Street Pier–Maryland data set

Description

Test data set from Stephens and Jennings SWProd calculator (USGS). The data has been interpolated to make it have readings every fifteen minutes.

Usage

data(DOTemp)

Format

DateTime

DateTime

Temp

Temperature in Celcius

DO

Dissolved Oxygen

Details

This is only for example and should be used judiciously for any kind of ecosystem interpretation (requires zoo and chron packages).

Source

Stephens, D.W., and Jennings, M.E., 1976, Determination of primary productivity and community metabolism in streams and lakes using diel oxygen measurements: U.S. Geological Survey Computer Contribution, 100 p.

[Package StreamMetabolism version 1.1.2 Index]