cfs.lps {StreamMetabolism}	R Documentation

Convert from cubic feet per second to liters per second

Description

Convert from cubic feet persecond to liters per second

Usage

cfs.lps(x)

Arguments

x
Discharge in cfs

Author(s)

Stephen A Sefick Jr.

[Package StreamMetabolism version 1.1.2 Index]