contiguous.zoo {StreamMetabolism}	R Documentation

contiguous.zoo

Description

find continuous non NA portions of zoo time series data

Usage

contiguous.zoo(x)

Arguments

x
zoo time series object whatever indexes you want

Details

if you want to just find the contiguous portions of just one signal and not the interaction between two just duplicate the signal contiguous.zoo(data.frame(x, coredata(x))) should give you what you want

Value

data frame constiting of

start
start index

end
end index

lengths
length of record

value
logical stating whether a continuous string of non-NA values

Author(s)

Gabor Grothendieck and Stephen A Sefick Jr.

[Package StreamMetabolism version 1.1.2 Index]