dC.dt {StreamMetabolism}	R Documentation

Change in Oxygen per time step

Description

Calculate the rate of change of Dissolved Oxygen

Usage

dC.dt(x)

Arguments

x
Dissolved Oxygen time series

Details

input zoo series takes the difference of DOt+1 - DOt

Value

zoo series of Dissolved Oxygen Differences with an NA for the first value as there is no value before that to subtract

Author(s)

Stephen A Sefick Jr.

References

Odum, H. T. (1956). "Primary production in flowing waters." Limnology and Oceanography, 1: 102-117.

Examples

data(DOTemp)
Diffconc <- dC.dt(DOTemp[,2])
plot(Diffconc)

[Package StreamMetabolism version 1.1.2 Index]