read.production {StreamMetabolism}	R Documentation

Read in Time Series Data as zoo Object

Description

Wrapper Function to read.zoo

Usage

read.production(data)

Arguments

data
a csv file with headers and the date as mm/dd/yyyy hh:mm:ss format (think excel spreadsheet date format and the file is saved as a csv file)

Details

This is a wrapper function to read.zoo with a specific format required see above

Value

zoo object

Author(s)

Stephen A Sefick Jr

See Also

read.table

[Package StreamMetabolism version 1.1.2 Index]